Refactoring model
