+++ account/src/main/java/pl/edu/pja/prz/account/model/Account.java
@@ -1,47 +1,25 @@
 package pl.edu.pja.prz.account.model;
 
 
-import pl.edu.pja.prz.account.model.value.Address;
-import pl.edu.pja.prz.account.model.value.FullName;
+import pl.edu.pja.prz.account.model.enums.AccountStatus;
 import pl.edu.pja.prz.account.model.value.Password;
-import pl.edu.pja.prz.account.model.value.Phone;
 
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.ManyToMany;
+import javax.persistence.*;
 import java.util.Set;
-import java.util.UUID;
 
 @Entity
-public abstract class Account extends BaseEntity<UUID> {
+@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
+public abstract class Account extends Person  {
 
-	private Phone phoneNumber;
 	private String email;
-	private FullName fullName;
-	private Address address;
 	private Password password;
-
+	@Enumerated(EnumType.STRING)
+	private AccountStatus accountStatus;
 	@ManyToMany(mappedBy = "accounts", fetch = FetchType.LAZY)
 	private Set<Role> roles;
 
 	Account() { }
 
-	public Account(Phone phoneNumber, String email, FullName fullName, Address address, Password password, Set<Role> roles) {
-		this.phoneNumber = phoneNumber;
-		this.email = email;
-		this.fullName = fullName;
-		this.address = address;
-		this.password = password;
-		this.roles = roles;
-	}
-
-	public Phone getPhoneNumber() {
-		return phoneNumber;
-	}
-
-	public void setPhoneNumber(Phone phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
 
 	public String getEmail() {
 		return email;
@@ -51,28 +29,20 @@ public abstract class Account extends BaseEntity<UUID> {
 		this.email = email;
 	}
 
-	public FullName getFullName() {
-		return fullName;
-	}
-
-	public void setFullName(FullName fullName) {
-		this.fullName = fullName;
-	}
-
-	public Address getAddress() {
-		return address;
-	}
-
-	public void setAddress(Address address) {
-		this.address = address;
-	}
-
 	public Password getPassword() {
 		return password;
 	}
 
 	public void setPassword(Password password) {
 		this.password = password;
+	}
+
+	public AccountStatus getAccountStatus() {
+		return accountStatus;
+	}
+
+	public void setAccountStatus(AccountStatus accountStatus) {
+		this.accountStatus = accountStatus;
 	}
 
 	public Set<Role> getRoles() {
@@ -86,7 +56,6 @@ public abstract class Account extends BaseEntity<UUID> {
 	public boolean addRole(Role role) {
 		role.getAccounts().add(this);
 		return this.roles.add(role);
-
 	}
 
 	public boolean removeRole(Role role) {
