+++ account/src/main/java/pl/edu/pja/prz/account/model/AccountAbstractFactory.java
@@ -1,22 +1,20 @@
 package pl.edu.pja.prz.account.model;
 
-import pl.edu.pja.prz.account.model.value.*;
-
 public interface AccountAbstractFactory {
 
-
-	Administrator createAdministrator(String email, Phone phone, Password password, FullName fullName,
-	                                  Address address);
-
-	Teacher createTeacher(String email, Phone phone, Password password, FullName fullName,
-	                      Address address, IdentityObject<Long>... groups);
-
-	Teacher createTeacher(String email, Phone phone, Password password, FullName fullName,
-	                      Address address);
-
-	Guardian createGuardian(String email, Phone phone, Password password, FullName fullName,
-	                        Address address, Child... children);
-
-	Guardian createGuardian(String email, Phone phone, Password password, FullName fullName,
-	                        Address address);
+//
+//	Administrator createAdministrator(String email, Phone phone, Password password, FullName fullName,
+//	                                  Address address);
+//
+//	Employee createTeacher(String email, Phone phone, Password password, FullName fullName,
+//	                       Address address, IdentityObject<Long>... groups);
+//
+//	Employee createTeacher(String email, Phone phone, Password password, FullName fullName,
+//	                       Address address);
+//
+//	Guardian createGuardian(String email, Phone phone, Password password, FullName fullName,
+//	                        Address address, Child... children);
+//
+//	Guardian createGuardian(String email, Phone phone, Password password, FullName fullName,
+//	                        Address address);
 }
