+++ account/src/main/java/pl/edu/pja/prz/account/model/BaseEntity.java
@@ -5,10 +5,8 @@ import javax.persistence.*;
 @MappedSuperclass
 public class BaseEntity<T> {
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
+	@Id @GeneratedValue(strategy = GenerationType.AUTO)
 	private T id;
-
 	@Version
 	private Integer version;
