+++ account/src/main/java/pl/edu/pja/prz/account/model/Person.java
@@ -1,0 +1,38 @@
+package pl.edu.pja.prz.account.model;
+
+import pl.edu.pja.prz.account.model.value.Address;
+import pl.edu.pja.prz.account.model.value.FullName;
+import pl.edu.pja.prz.account.model.value.Phone;
+
+import java.util.UUID;
+
+public class Person extends BaseEntity<UUID>{
+
+	private Phone phoneNumber;
+	private FullName fullName;
+	private Address address;
+
+	public Phone getPhoneNumber() {
+		return phoneNumber;
+	}
+
+	public void setPhoneNumber(Phone phoneNumber) {
+		this.phoneNumber = phoneNumber;
+	}
+
+	public FullName getFullName() {
+		return fullName;
+	}
+
+	public void setFullName(FullName fullName) {
+		this.fullName = fullName;
+	}
+
+	public Address getAddress() {
+		return address;
+	}
+
+	public void setAddress(Address address) {
+		this.address = address;
+	}
+}
