+++ account/src/main/java/pl/edu/pja/prz/account/model/enums/AccountStatus.java
@@ -1,0 +1,5 @@
+package pl.edu.pja.prz.account.model.enums;
+
+public enum AccountStatus {
+	ACCTIVE,
+}
